requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
httpx==0.27.0
python-dateutil==2.9.0.post0
google-api-python-client==2.126.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
pytest==8.3.3
